@@ -3,7 +3,7 @@ name = "advent-of-code-2025-day09"
 version = "0.1.0"
 description = "Advent of Code 2025 Day 09"
 authors = [{name = "Judson Powers"}]
-requires-python = ">=3.8"
+requires-python = ">=3.12"
 
 [project.optional-dependencies]
 dev = ["pytest", "ruff", "mypy"]
